% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{unit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:unit]{unit}}}
}}

